+++ commands/AdminController.php
@@ -2,7 +2,7 @@
 
 namespace app\commands;
 
-use app\components\SitePurger;
+use app\components\{IMotionStatusFilter, SitePurger};
 use app\models\db\{Amendment, Consultation, Motion, Site};
 use app\models\settings\AntragsgruenApp;
 use yii\console\Controller;
@@ -73,7 +73,7 @@ class AdminController extends Controller
             return;
         }
 
-        foreach ($con->getVisibleMotions() as $motion) {
+        foreach (IMotionStatusFilter::onlyUserVisible($con, true)->getFilteredConsultationMotions() as $motion) {
             echo '- Motion ' . $motion->id . "\n";
             $motion->getNumberOfCountableLines();
             $motion->getFirstLineNumber();