+++ controllers/AmendmentController.php
@@ -13,7 +13,7 @@ use app\models\http\{BinaryFileResponse,
     ResponseInterface,
     RestApiExceptionResponse,
     RestApiResponse};
-use app\components\{HTMLTools, Tools, UrlHelper};
+use app\components\{HTMLTools, IMotionStatusFilter, Tools, UrlHelper};
 use app\models\db\{Amendment, AmendmentAdminComment, AmendmentSupporter, ConsultationLog, ISupporter, Motion, User};
 use app\models\events\AmendmentEvent;
 use app\models\exceptions\{FormError, MailNotSent, ResponseException};
@@ -65,14 +65,15 @@ class AmendmentController extends Base
         );
     }
 
-    public function actionPdfcollection(int $withdrawn = 0): ResponseInterface
+    public function actionPdfcollection(int $inactive = 0): ResponseInterface
     {
-        $withdrawn = ($withdrawn === 1);
-        $motions   = $this->consultation->getVisibleIMotionsSorted($withdrawn);
+        $filter = IMotionStatusFilter::adminExport($this->consultation, ($inactive === 1));
+        $motions = $filter->getFilteredConsultationIMotionsSorted();
         if (count($motions) === 0) {
             return new HtmlErrorResponse(404, \Yii::t('motion', 'none_yet'));
         }
-        $amendments  = [];
+        $amendments = [];
+        $motion = null;
         $texTemplate = null;
         foreach ($motions as $motion) {
             if (!is_a($motion, Motion::class) || $motion->getMyMotionType()->amendmentsOnly) {
@@ -82,9 +83,9 @@ class AmendmentController extends Base
             if ($texTemplate === null) {
                 $texTemplate = $motion->getMyMotionType()->texTemplate;
             }
-            $amendments = array_merge($amendments, $motion->getVisibleAmendmentsSorted($withdrawn));
+            $amendments = array_merge($amendments, $motion->getFilteredAmendments($filter));
         }
-        if (count($amendments) == 0) {
+        if (count($amendments) === 0) {
             return new HtmlErrorResponse(404, \Yii::t('amend', 'none_yet'));
         }