+++ controllers/MotionExportTraits.php
@@ -4,7 +4,7 @@ namespace app\controllers;
 
 use app\views\pdfLayouts\IPDFLayout;
 use app\models\settings\{PrivilegeQueryContext, Privileges, AntragsgruenApp};
-use app\components\{RequestContext, Tools, UrlHelper};
+use app\components\{IMotionStatusFilter, RequestContext, Tools, UrlHelper};
 use app\models\exceptions\NotFound;
 use app\models\http\{BinaryFileResponse,
     HtmlErrorResponse,
@@ -163,11 +163,13 @@ trait MotionExportTraits
         );
     }
 
-    private function getMotionsAndTemplate(string $motionTypeId, bool $withdrawn, bool $resolutions): array
+    private function getMotionsAndTemplate(string $motionTypeId, bool $inactive, bool $resolutions): array
     {
         /** @var TexTemplate $texTemplate */
         $texTemplate = null;
-        $imotions = $this->consultation->getVisibleIMotionsSorted($withdrawn);
+        $imotions = IMotionStatusFilter::adminExport($this->consultation, $inactive)
+                                       ->getFilteredConsultationIMotionsSorted();
+
         if ($motionTypeId !== '' && $motionTypeId !== '0') {
             $motionTypeIds = explode(',', $motionTypeId);
             $imotions       = array_filter($imotions, function (IMotion $motion) use ($motionTypeIds) {
@@ -197,13 +199,10 @@ trait MotionExportTraits
         return [$imotionsFiltered, $texTemplate];
     }
 
-    public function actionFullpdf(string $motionTypeId = '', int $withdrawn = 0, int $resolutions = 0): ResponseInterface
+    public function actionFullpdf(string $motionTypeId = '', int $inactive = 0, int $resolutions = 0): ResponseInterface
     {
-        $withdrawn = ($withdrawn === 1);
-        $resolutions = ($resolutions === 1);
-
         try {
-            list($imotions, $texTemplate) = $this->getMotionsAndTemplate($motionTypeId, $withdrawn, $resolutions);
+            list($imotions, $texTemplate) = $this->getMotionsAndTemplate($motionTypeId, ($inactive === 1), ($resolutions === 1));
             /** @var IMotion[] $imotions */
             if (count($imotions) === 0) {
                 return new HtmlErrorResponse(404, \Yii::t('motion', 'none_yet'));
@@ -236,13 +235,10 @@ trait MotionExportTraits
         );
     }
 
-    public function actionPdfcollection(string $motionTypeId = '', int $withdrawn = 0, int $resolutions = 0): ResponseInterface
+    public function actionPdfcollection(string $motionTypeId = '', int $inactive = 0, int $resolutions = 0): ResponseInterface
     {
-        $withdrawn = ($withdrawn === 1);
-        $resolutions = ($resolutions === 1);
-
         try {
-            list($imotions, $texTemplate) = $this->getMotionsAndTemplate($motionTypeId, $withdrawn, $resolutions);
+            list($imotions, $texTemplate) = $this->getMotionsAndTemplate($motionTypeId, ($inactive === 1), ($resolutions === 1));
             if (count($imotions) === 0) {
                 return new HtmlErrorResponse(404, \Yii::t('motion', 'none_yet'));
             }
@@ -252,7 +248,8 @@ trait MotionExportTraits
                 $imotions = [];
                 foreach ($motionType->motions as $motion) {
                     if (is_a($motion, Motion::class)) {
-                        $imotions = array_merge($imotions, $motion->getVisibleAmendmentsSorted($withdrawn));
+                        $filter = IMotionStatusFilter::adminExport($this->consultation, ($inactive === 1));
+                        $imotions = array_merge($imotions, $motion->getFilteredAndSortedAmendments($filter));
                     }
                 }
                 if (count($imotions) === 0) {